Update DuolingoTreeEnhancer.user.js
I just started in reverse mode, I added so hide and speak were checked on both languages. 
It was showing the orange new words in English (in my English to Spanish reverse tree)
e.g. Rice and beans, and would be hidden but not rice and beans. 
I added the classes so that both rice and beans are hidden too.